org: moved argument types to Parse.h
