+++ src/Types.h
@@ -7,9 +7,6 @@
 #ifndef TYPES_H_
 #define TYPES_H_
 
-#include "charm++.h"
-#include "pup_stl.h"
-
 #include <cstdint>
 #include <string>
 
@@ -64,59 +61,6 @@ struct Grid {
       && 0 != rhs.height
       && (0 == width % rhs.width)
       && (0 == height % rhs.height);
-  }
-};
-
-struct OnTheFlyArguments {
-  Grid<Id> personGrid;
-  Grid<Id> locationGrid;
-  Grid<PartitionId> locationPartitionGrid;
-  Grid<Id> localLocationGrid;
-
-  Id averageVisitsPerDay;
-};
-PUPbytes(OnTheFlyArguments);
-
-struct Arguments {
-  // Arguments needed for all runs
-  PartitionId numPersonPartitions;
-  PartitionId numLocationPartitions;
-  Time numDays;
-  Time numDaysWithDistinctVisits;
-  Time numDaysToSeedOutbreak;
-  Id numInitialInfectionsPerDay;
-  int seed;
-
-  bool hasIntervention;
-  int contactModelType;
-
-  std::string diseasePath;
-  std::string interventionPath;
-  std::string outputPath;
-
-  bool isOnTheFlyRun;
-  struct OnTheFlyArguments onTheFly;
-  std::string scenarioPath;
-
-  Arguments() {}
-  explicit Arguments(CkMigrateMessage *msg) {}
-
-  void pup(PUP::er &p) {
-    p | numPersonPartitions;
-    p | numLocationPartitions;
-    p | numDays;
-    p | numDaysWithDistinctVisits;
-    p | numDaysToSeedOutbreak;
-    p | numInitialInfectionsPerDay;
-    p | seed;
-    p | hasIntervention;
-    p | contactModelType;
-    p | diseasePath;
-    p | interventionPath;
-    p | outputPath;
-    p | scenarioPath;
-    p | isOnTheFlyRun;
-    p | onTheFly;
   }
 };
