+++ src/loimos.ci
@@ -7,6 +7,7 @@ mainmodule loimos {
   include "Types.h";
   include "Interaction.h";
   include "Message.h";
+  include "readers/Parse.h";
 
 #ifdef USE_HYPERCOMM
   include "AggregatorParam.h";
